Move FakeControllers to Fixtures
